pyproject.toml
[tool.poetry]
name = "image-text-retrieval"
version = "0.1.0"
description = "A cool repo to experiment with image and text retrieval"
authors = ["paluchasz <paluchasz@googlemail.com>"]
readme = "readme.md"
packages = [{ include = "image_text_retrieval" }]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.1"
pydantic = "^2.6.4"
transformers = "^4.38.2"
numpy = "^1.26.4"
loguru = "^0.7.2"
datasets = "^2.18.0"
pillow = "^10.2.0"
pandas = "^2.2.1"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
pydantic-settings = "^2.2.1"
more-itertools = "^10.2.0"
scikit-learn = "^1.4.1.post1"
scipy = "^1.12.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.3"
mypy = "^1.9.0"
black = "^24.3.0"
pre-commit = "^3.6.2"
types-pillow = "^10.2.0.20240311"

[tool.poetry.scripts]
image_text_retrieval_api = "image_text_retrieval.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 150

[tool.mypy]
plugins = ["pydantic.mypy"]
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
show_column_numbers = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[[tool.mypy.overrides]]
module = ["transformers", "datasets", "pandas", "tqdm", "scipy.special"]
ignore_missing_imports = true

[tool.ruff]
extend-select = [
    "B",
    "BLE",
    "C4",
    "ERA",
    "I",
    "ICN",
    "INP",
    "ISC",
    "N",
    "NPY",
    "PGH",
    "PIE",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "T20",
    "TCH",
    "TID",
    "YTT",
]
ignore = ["E402", "B905", "N818", "S101", "S104"]
show-fixes = true
line-length = 150
target-version = "py39"
